+++ changelog.md
@@ -6,6 +6,7 @@
 - Batch processing plan fixes: 
    - Fix adding pipelines to plan
    - Redesign mask widget
+- modify measurement backend to allow calculate multi channel measurements. 
 
 ## 0.9
 Begin of changelog